requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
pytz>=2021.1
tabulate>=0.8.9
colorama>=0.4.4
tqdm>=4.62.0
python-dateutil>=2.8.2